+++ R/match_md_tables.R
@@ -3,7 +3,7 @@
 #' @details `match_md_tables` attempts to match all markdown
 #'   tables in `file` utilizing a regular expression
 #'
-#' @param content Sourced raw markdown from `source_file`
+#' @param content Sourced raw markdown from `readMDTable::source_file`
 #'
 #' @returns `NULL` if no markdown table content is found,
 #'   else a list of table matches.
@@ -13,9 +13,9 @@
 #' @noRd
 match_md_tables <- function(content) {
   table_pattern <- paste0(
-    "\\|(?:([^\r\n|]*)\\|)+\r?\n",                        # Header
-    "\\|\\s*(:?-+:?)\\s*(\\|\\s*(:?-+:?)\\s*)*\\|?\r?\n", # Separator
-    "(\\|(?:([^\\r\\n|]*)\\|)+(\\r?\\n)?)+"               # Data
+    "\\|(?:([^\r\n|]*)\\|)+\r?\n",          # Header
+    "(\\|\\s*(:?-+:?)\\s*)*\\|?\r?\n",      # Separator
+    "(\\|(?:([^\\r\\n|]*)\\|)+(\\r?\\n)?)+" # Data
   )
 
   table_matches <- gregexpr(table_pattern, content, perl = TRUE)